fix: restore required usings
